requirements.txt
flask
google-generativeai>=0.5.4
python-dotenv
pypdf
gunicorn